op-service,op-node: move three test helpers out of production packages
Same class as the parent commit: files that are test support but are not named
_test.go, so they compile as production code and put test machinery into the
build closure of everything importing them.

op-service/metrics/checker.go defines MetricFamilyChecker, a test assertion
helper, in a package nearly every OP binary imports. It is the reason
stretchr/testify, objx, go-spew and go-difflib link into 35 of them. Move it to
op-service/metrics/test, matching op-node/rollup/derive/test; its seven users
are all _test.go files.

op-service/testutils/fake_publish.go has one user, op-e2e's l2_verifier. It is
the only file pulling op-service/apis into testutils, and with it libp2p, pion,
btcsuite and pebble: moving it to op-e2e drops that package's own closure from
803 to 725 packages and 91 to 72 heavy ones.

op-node/rollup/attributes/testutils.go defines a mock used only by
attributes_test.go in the same package, so it is a plain rename.

testify in shipped binaries: 35 -> 26.
