+++ op-e2e/actions/helpers/fake_publish.go
@@ -1,4 +1,4 @@
-package testutils
+package helpers
 
 import (
 	"context"
